# BoleteriaOnline.Core/ViewModels/Responses/DistribucionResponse.cs
namespace BoleteriaOnline.Web.ViewModels.Responses;
public class DistribucionResponse
{
    public int Id { get; set; }
    public List<FilaResponse> Filas { get; set; }
    public string Nota { get; set; }
    public bool UnPiso { get; set; }
}