# backend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Boleteria Online",
    "description": "Colleción de API's correspondientes al sistema de Pasajes Online.",
    "contact": {
      "name": "Agustin Ibañez",
      "email": "admin@boleteriasystem.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/distribuciones": {
      "get": {
        "tags": [
          "Distribuciones"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseICollectionWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseICollectionWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseICollectionWebResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Distribuciones"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DistribucionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DistribucionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DistribucionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Distribuciones"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/distribuciones/{id}": {
      "get": {
        "tags": [
          "Distribuciones"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Distribuciones"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DistribucionUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DistribucionUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DistribucionUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/distribuciones/{id}/filas": {
      "post": {
        "tags": [
          "Distribuciones"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Planta"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Planta"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Planta"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribucionResponseWebResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CeldaRequest": {
        "type": "object",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/DistribucionEspacio"
          }
        },
        "additionalProperties": false
      },
      "CeldaResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "$ref": "#/components/schemas/DistribucionEspacio"
          }
        },
        "additionalProperties": false
      },
      "CeldaUpdateRequest": {
        "type": "object",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/DistribucionEspacio"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "DistribucionEspacio": {
        "enum": [
          1,
          2,
          3,
          -1
        ],
        "type": "integer",
        "format": "int32"
      },
      "DistribucionRequest": {
        "required": [
          "filas",
          "nota",
          "un_piso"
        ],
        "type": "object",
        "properties": {
          "filas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilaRequest"
            }
          },
          "nota": {
            "maxLength": 128,
            "type": "string"
          },
          "un_piso": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "DistribucionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "filas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilaResponse"
            },
            "nullable": true
          },
          "nota": {
            "type": "string",
            "nullable": true
          },
          "un_piso": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "DistribucionResponseICollectionWebResult": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DistribucionResponse"
            },
            "nullable": true
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/ErrorMessage"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DistribucionResponseWebResult": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/DistribucionResponse"
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/ErrorMessage"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DistribucionUpdateRequest": {
        "required": [
          "filas",
          "nota",
          "un_piso"
        ],
        "type": "object",
        "properties": {
          "filas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilaUpdateRequest"
            }
          },
          "nota": {
            "maxLength": 128,
            "type": "string"
          },
          "un_piso": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ErrorMessage": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11
        ],
        "type": "integer",
        "format": "int32"
      },
      "FilaRequest": {
        "required": [
          "cells",
          "planta"
        ],
        "type": "object",
        "properties": {
          "cells": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CeldaRequest"
            }
          },
          "planta": {
            "$ref": "#/components/schemas/Planta"
          }
        },
        "additionalProperties": false
      },
      "FilaResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "cells": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CeldaResponse"
            },
            "nullable": true
          },
          "planta": {
            "$ref": "#/components/schemas/Planta"
          }
        },
        "additionalProperties": false
      },
      "FilaUpdateRequest": {
        "required": [
          "cells",
          "id",
          "planta"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "cells": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CeldaUpdateRequest"
            }
          },
          "planta": {
            "$ref": "#/components/schemas/Planta"
          }
        },
        "additionalProperties": false
      },
...
      "Planta": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}